+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next version
 - Remove `GetHTTP`
+- I18n fixes
 
 ## 1.5.13
 - Fix `info` helper with nested blank value