+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Remove `GetHTTP`
 - I18n fixes
 - Add `Agilibox::TapMethods`
+- Filters CSS : add flex-wrap
 
 ## 1.5.13
 - Fix `info` helper with nested blank value